requirements.txt
# Core web framework
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1

# Configuration
PyYAML==6.0.2
python-dotenv==1.1.1

# Document processing
pypdf==6.0.0
PyPDF2==3.0.1
python-docx==1.2.0
beautifulsoup4==4.13.5
html5lib==1.1
html2text==2025.4.15
lxml==6.0.1

# Embeddings and vector operations
sentence-transformers==5.1.0
numpy>=2.3.2
scipy>=1.16.0
nltk==3.9.1

# Vector stores
chromadb==1.0.20
qdrant-client==1.15.1

# AI/LLM providers
openai==1.3.7
google-generativeai==0.3.2

# Retrieval and evaluation
scikit-learn>=1.7.1
tqdm==4.67.1

# HTTP client for external APIs
httpx==0.28.1

# Better logging (optional)
loguru==0.7.3

# Development and testing
pytest==8.4.1
pytest-asyncio==1.1.0

python-multipart==0.0.20